Added CommentResponse

+++ src/main/java/dev/accessaid/AccessAid/Comments/controller/CommentController.java
@@ -3,6 +3,8 @@ package dev.accessaid.AccessAid.Comments.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -12,8 +14,11 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import dev.accessaid.AccessAid.Comments.Response.CommentResponse;
 import dev.accessaid.AccessAid.Comments.model.Comment;
 import dev.accessaid.AccessAid.Comments.service.CommentService;
+import dev.accessaid.AccessAid.Comments.utils.CommentMapper;
+import dev.accessaid.AccessAid.Geolocation.Response.ErrorResponse;
 import dev.accessaid.AccessAid.Places.model.Place;
 import dev.accessaid.AccessAid.model.User;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -27,42 +32,81 @@ public class CommentController {
     private CommentService commentService;
 
     @GetMapping("")
-    public List<Comment> seeAllComments() {
-        return commentService.getComments();
+    public List<CommentResponse> seeAllComments() {
+        List<Comment> comments = commentService.getComments();
+        return CommentMapper.toCommentResponses(comments);
     }
 
     @GetMapping("/{id}")
-    public Comment seeCommentById(@PathVariable Integer id) {
-        return commentService.getCommentById(id);
+    public CommentResponse seeCommentById(@PathVariable Integer id) {
+        Comment comment = commentService.getCommentById(id);
+        return CommentMapper.toCommentResponse(comment);
+
     }
 
     @PostMapping("")
-    public Comment addComment(@RequestBody Comment comment) {
-
-        return commentService.createComment(comment);
+    public ResponseEntity<?> addComment(@RequestBody Comment comment) {
+        try {
+            Comment createdComment = commentService.createComment(comment);
+            CommentResponse response = CommentMapper.toCommentResponse(createdComment);
+            return ResponseEntity.ok(response);
+        } catch (Exception e) {
+            System.out.println(e);
+            ErrorResponse errorResponse = new ErrorResponse("Error");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
+        }
     }
 
     @PutMapping("/{id}")
-    public Comment updateComment(@PathVariable Integer id, @RequestBody Comment comment) {
-        return commentService.changeComment(comment);
+    public ResponseEntity<?> updateComment(@PathVariable Integer id, @RequestBody Comment comment) {
+        try {
+            Comment existingComment = commentService.getCommentById(id);
+            if (existingComment == null) {
+                ErrorResponse errorResponse = new ErrorResponse("Comment not found with id: " + id);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+            }
+            existingComment.setComment(comment.getComment());
+            Comment updatedComment = commentService.changeComment(existingComment);
+            CommentResponse response = CommentMapper.toCommentResponse(updatedComment);
+            return ResponseEntity.ok(response);
+        } catch (Exception e) {
+            System.out.println(e);
+            ErrorResponse errorResponse = new ErrorResponse("Error");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public Comment deleteComment(@PathVariable Integer id) {
-        return commentService.removeComment(id);
+    public ResponseEntity<?> deleteComment(@PathVariable Integer id) {
+        try {
+            Comment commentToDelete = commentService.getCommentById(id);
+            if (commentToDelete == null) {
+                ErrorResponse errorResponse = new ErrorResponse("Comment not found with id: " + id);
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+            }
+            commentService.removeComment(id);
+            CommentResponse response = CommentMapper.toCommentResponse(commentToDelete);
+            return ResponseEntity.ok(response);
+        } catch (Exception e) {
+            System.out.println(e);
+            ErrorResponse errorResponse = new ErrorResponse("Error");
+            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body(errorResponse);
+        }
     }
 
     @GetMapping("/place/{placeId}")
-    public List<Comment> seeCommentsByPlace(@PathVariable Integer placeId) {
+    public List<CommentResponse> seeCommentsByPlace(@PathVariable Integer placeId) {
         Place place = new Place();
         place.setId(placeId);
-        return commentService.getAllCommentsByPlace(place);
+        List<Comment> comments = commentService.getAllCommentsByPlace(place);
+        return CommentMapper.toCommentResponses(comments);
     }
 
     @GetMapping("/user/{userId}")
-    public List<Comment> seeCommentsByUser(@PathVariable Integer userId) {
+    public List<CommentResponse> seeCommentsByUser(@PathVariable Integer userId) {
         User user = new User();
         user.setId(userId);
-        return commentService.getAllCommentsByUser(user);
+        List<Comment> comments = commentService.getAllCommentsByUser(user);
+        return CommentMapper.toCommentResponses(comments);
     }
 }
